requirement.txt
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
pkg-resources==0.0.0
protobuf==3.6.1
six==1.14.0
Werkzeug==1.0.1
